tv5725: one class per file for the output raster
RasterSolution added at 134 and renamed away at 136.

Folded, in order:

    tv5725: name the output timings after what owns them
    tv5725: say what clockDividerFor returns, not how it chooses

+++ GBSC-Pro-Source code/gbs-control/src/tv5725/BlankingTiming.cpp
@@ -4,13 +4,13 @@ namespace Tv5725 {
 
 BlankingTiming::BlankingTiming() : stop_(0), start_(0) {}
 
-BlankingTiming::BlankingTiming(uint16_t stop, uint16_t start) : stop_(stop), start_(start) {}
+BlankingTiming::BlankingTiming(int32_t stop, int32_t start) : stop_(stop), start_(start) {}
 
-uint16_t BlankingTiming::stop() const { return stop_; }
+int32_t BlankingTiming::stop() const { return stop_; }
 
-uint16_t BlankingTiming::start() const { return start_; }
+int32_t BlankingTiming::start() const { return start_; }
 
-uint16_t BlankingTiming::width() const { return start_ > stop_ ? start_ - stop_ : 0; }
+int32_t BlankingTiming::width() const { return start_ > stop_ ? start_ - stop_ : 0; }
 
 bool BlankingTiming::usable() const { return start_ > stop_; }
